♻️ refactor: 폴더 구조 피드백 반영 수정
