+++ src/features/comment/ui/add-comment-button/index.tsx
@@ -1,4 +1,3 @@
-
 import { Button } from "@shared/ui"
 import { Comment } from "@entities/comment/model"
 import { ThumbsUp, Edit2, Trash2 } from "lucide-react"
