--- src/features/comment/ui/comment-form/index.ts
@@ -1,1 +1,0 @@
-export * from "./CommentForm"
