--- src/features/comment/ui/comment-list/index.ts
@@ -1,2 +1,0 @@
-export * from "./CommentsList"
-export * from "./PostCommentsFeature"
