--- src/features/comment/ui/comment-management/index.ts
@@ -1,1 +1,0 @@
-export * from "./CommentActionButtons"
