+++ src/features/post/ui/index.ts
@@ -1,4 +1,4 @@
-export * from "./post-filtering"
-export * from "./post-form"
-export * from "./post-management"
-export * from "./post-search"
+export { PostFilter } from "./post-filter"
+export { PostForm } from "./post-form"
+export { PostActionButtons } from "./add-post-button"
+export { PostSearch } from "./post-search"
