--- src/features/post/ui/post-filtering/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterControls"
