--- src/features/post/ui/post-form/index.ts
@@ -1,1 +1,0 @@
-export * from "./PostFormFields"
