--- src/features/post/ui/post-management/index.ts
@@ -1,1 +1,0 @@
-export * from "./PostActionButtons"
