--- src/features/post/ui/post-search/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchInput"
