--- src/features/user/ui/user-info/index.ts
@@ -1,1 +1,0 @@
-export * from "./UserInfo"
