+++ src/widgets/comment/ui/index.ts
@@ -1,1 +1,1 @@
-export * from "./CommentFormDialog"
+export { CommentFormDialog } from "./CommentFormDialog"
