+++ src/widgets/post/ui/PostsSearchFilter.tsx
@@ -1,5 +1,5 @@
-import { SearchInput } from "@features/post/ui"
-import { FilterControls } from "@features/post/ui"
+import { PostSearch } from "@features/post/ui"
+import { PostFilter } from "@features/post/ui"
 import { PostTag } from "@entities/post/model"
 
 interface PostsSearchFilterProps {
@@ -29,9 +29,9 @@ export const PostsSearchFilter = ({
 }: PostsSearchFilterProps) => (
   <div className="flex gap-4">
     <div className="flex-1">
-      <SearchInput value={searchQuery} onChange={onSearchChange} onSearch={onSearch} />
+      <PostSearch value={searchQuery} onChange={onSearchChange} onSearch={onSearch} />
     </div>
-    <FilterControls
+    <PostFilter
       selectedTag={selectedTag}
       tags={tags}
       onTagChange={onTagChange}
