+++ src/widgets/post/ui/index.ts
@@ -1,5 +1,5 @@
-export * from "./PostDetailDialog"
-export * from "./PostFormDialog"
-export * from "./PostsManagerWidget"
-export * from "./PostsSearchFilter"
-export * from "./PostsTable"
+export { PostDetailDialog } from "./PostDetailDialog"
+export { PostFormDialog } from "./PostFormDialog"
+export { PostsManagerWidget } from "./PostsManagerWidget"
+export { PostsSearchFilter } from "./PostsSearchFilter"
+export { PostsTable } from "./PostsTable"
