+++ src/widgets/user/ui/index.ts
@@ -1,1 +1,1 @@
-export * from "./UserInfoDialog"
+export { UserInfoDialog } from "./UserInfoDialog"
