+++ tests/basic.test.tsx
@@ -68,8 +68,8 @@ describe("PostsManager", () => {
     })
 
     // 검색 기능 테스트
-    const searchInput = screen.getByPlaceholderText(/게시물 검색.../i)
-    await user.type(searchInput, "His mother had always taught him")
+    const PostSearch = screen.getByPlaceholderText(/게시물 검색.../i)
+    await user.type(PostSearch, "His mother had always taught him")
     await user.keyboard("{Enter}")
 
     await waitFor(() => {
